chore: fix examples
