--- not_examples/default.rs
@@ -1,51 +1,0 @@
-use std::path::PathBuf;
-use zk_engine::{
-  args::{WASMArgsBuilder, WASMCtx},
-  // Backend imports
-  nova::{
-    provider::{ipa_pc, PallasEngine},
-    spartan::{self, snark::RelaxedR1CSSNARK},
-    traits::Dual,
-  },
-  run::default::ZKEProof,
-  traits::zkvm::ZKVM,
-  utils::logging::init_logger,
-};
-
-// Curve cycle to use for proving
-type E1 = PallasEngine;
-// PCS used for final SNARK at the end of (N)IVC
-type EE1<E> = ipa_pc::EvaluationEngine<E>;
-// PCS on secondary curve
-type EE2<E> = ipa_pc::EvaluationEngine<Dual<E>>;
-
-// Spartan SNARKS used for compressing at then end of (N)IVC
-type BS1<E> = spartan::batched::BatchedRelaxedR1CSSNARK<E, EE1<E>>;
-type S1<E> = RelaxedR1CSSNARK<E, EE1<E>>;
-type S2<E> = RelaxedR1CSSNARK<Dual<E>, EE2<E>>;
-
-fn main() -> anyhow::Result<()> {
-  init_logger();
-
-  // Configure the arguments needed for WASM execution
-  //
-  // Here we are configuring the path to the WASM file
-  let args = WASMArgsBuilder::default()
-    .file_path(PathBuf::from("wasm/example.wasm"))
-    .build();
-
-  // Create a WASM execution context for proving.
-  let mut wasm_ctx = WASMCtx::new_from_file(args)?;
-
-  // Prove execution and run memory consistency checks
-  //
-  // Get proof for verification and corresponding public values
-  //
-  // Above type alias's (for the backend config) get used here
-  let (proof, public_values, _) =
-    ZKEProof::<E1, BS1<E1>, S1<E1>, S2<E1>>::prove_wasm(&mut wasm_ctx)?;
-
-  // Verify proof
-  let result = proof.verify(public_values)?;
-  Ok(assert!(result))
-}
