--- not_examples/zk_ads.rs
@@ -1,52 +1,0 @@
-use std::path::PathBuf;
-
-use zk_engine::{
-  args::{WASMArgsBuilder, WASMCtx},
-  nova::{
-    provider::{ipa_pc, PallasEngine},
-    spartan::{self, snark::RelaxedR1CSSNARK},
-    traits::Dual,
-  },
-  run::batched::BatchedZKEProof,
-  traits::zkvm::ZKVM,
-  utils::logging::init_logger,
-};
-
-// Curve cycle to use for proving
-type E1 = PallasEngine;
-// PCS used for final SNARK at the end of (N)IVC
-type EE1<E> = ipa_pc::EvaluationEngine<E>;
-// PCS on secondary curve
-type EE2<E> = ipa_pc::EvaluationEngine<Dual<E>>;
-
-// Spartan SNARKS used for compressing at then end of (N)IVC
-type BS1<E> = spartan::batched::BatchedRelaxedR1CSSNARK<E, EE1<E>>;
-type S1<E> = RelaxedR1CSSNARK<E, EE1<E>>;
-type S2<E> = RelaxedR1CSSNARK<Dual<E>, EE2<E>>;
-
-fn main() -> anyhow::Result<()> {
-  init_logger();
-
-  let input_x = "200.05";
-  let input_y = "-30.0";
-
-  let args = WASMArgsBuilder::default()
-    .file_path(PathBuf::from("wasm/zk_ads.wasm"))
-    .invoke(Some(String::from("is_user_close_enough")))
-    .func_args(vec![
-      String::from("0"),
-      String::from(input_x),
-      String::from(input_y),
-    ])
-    .build();
-
-  // Create a WASM execution context for proving.
-  let mut wasm_ctx = WASMCtx::new_from_file(args)?;
-
-  let (proof, public_values, _) =
-    BatchedZKEProof::<E1, BS1<E1>, S1<E1>, S2<E1>>::prove_wasm(&mut wasm_ctx)?;
-
-  // Verify proof
-  let result = proof.verify(public_values)?;
-  Ok(assert!(result))
-}
